requirements.txt
nose==1.3.7
coverage==4.4.1
pillow==4.2.0
flask==0.12.2
flask-restful==0.3.6
werkzeug==0.12.2
mock==2.0.0
opencv-python==3.2.0.7
numpy==1.13.0
scikit-learn==0.18.2
scipy==0.19.1
